backend/requirements.txt
# requirements.txt

asgiref==3.8.1
Brotli==1.1.0
cffi==1.17.1
channels==4.1.0
chardet==5.2.0
colorama==0.4.6
cssselect2==0.7.0
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
fonttools==4.54.1
gevent==24.11.1
greenlet==3.1.1
lxml==5.3.0
pillow==10.4.0
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.15.0
python-docx==1.1.2
python-dotenv==1.0.1
qrcode==8.0
reportlab==4.2.5
setuptools==75.4.0
sqlparse==0.5.1
tinycss2==1.4.0
tinyhtml5==2.0.0
typing_extensions==4.12.2
tzdata==2024.2
weasyprint==63.0
webencodings==0.5.1
websocket==0.2.1
whitenoise==6.8.2
zope.event==5.0
zope.interface==7.1.1
zopfli==0.2.3.post1
